Revert "feature2 changes are done (#1)"
This reverts commit 8c56a50481fefa7839c75abe0fa803b97d407772.
