+++ src/componnents/Body.css
@@ -10,15 +10,6 @@
 
 }
 
-.shimmer-card {
-    border: 1px solid black;
-    width : 195px;
-    margin: 10px;
-    padding : 10px;
-    display: flex;
-    flex-direction: column;
-}
-
 .cuisines {
     overflow-wrap :break-word;
 }
@@ -44,9 +35,4 @@
     font-size: 20px;;
     letter-spacing: 2px;
     border-radius: 10px;
-}
-
-.internetConError {
-
-    text-align: center;
 }
