+++ src/componnents/NavItems.js
@@ -1,11 +1,7 @@
 import React, { useState } from 'react'
 import { NavLink } from 'react-router-dom'
-import CircleIcon from '@mui/icons-material/Circle';
-import useOnline from '../utils/useOnline';
-
 export const NavItems = () => {
-  const [isLoggedIn,setIsLoggedIn] = useState(true);
-  const isOnline = useOnline()
+  const [isLoggedIn,setIsLoggedIn] = useState(true)
 
   return (
     <div className="nav-items">
@@ -29,10 +25,7 @@ export const NavItems = () => {
                 return setIsLoggedIn(!isLoggedIn)
                }}>LogIn</button>}
             </li>
-            {isOnline ?<CircleIcon fontSize='1rem' color="success"/> :
-             <CircleIcon fontSize='1rem' color="error"/>}
         </ul>
-        
     </div>
   )
 }
