--- src/pages/RestaurentMenu.css
@@ -1,9 +1,0 @@
-.resInfoImg {
-    width: 300px;
-    height: 300px;
-}
-
-.resInfo {
-    margin: 10px;
-    text-align: center;
-}
