--- src/utils/SearchInput.js
@@ -1,10 +1,0 @@
-const SearchInput = ({inputVal,onChangeFun}) => {
-    return (
-        <div className="input-search">
-            <input type="text" placeholder='Search Restaurent'
-            value={inputVal} onChange={onChangeFun}/>
-      </div>
-    )
-}
-
-export default SearchInput;
