--- src/utils/helper.js
@@ -1,5 +1,0 @@
-export  const filterdata = (inputVal,allRestaurants) => {
-    return allRestaurants.filter((restaurant)=>{
-      return restaurant?.data?.name?.toLowerCase()?.includes(inputVal?.toLowerCase())
-    })
-  }
