--- src/utils/useOnline.js
@@ -1,26 +1,0 @@
-import { useEffect,useState } from "react";
-
-const useOnline = () => {
-    const [isOnline,setIsOnline] = useState(true);
-
-    const handleOnline = ()=>{
-        setIsOnline(true)
-    }
-
-    const handleOffline = ()=>{
-        setIsOnline(false)
-    }
-    useEffect(() => {
-        window.addEventListener("online",handleOnline);
-        window.addEventListener("offline",handleOffline)
-        
-        return () => {
-            window.removeEventListener('online',handleOnline)
-            window.removeEventListener("offline",handleOffline)
-        }
-    }, []);
-
-    return isOnline
-}
-
-export default useOnline;
